feat(nextjs-example): Migrate to /app directory
