+++ examples/nextjs-example/app/globals.css
@@ -77,6 +77,7 @@
   box-sizing: border-box;
   padding: 0;
   margin: 0;
+  line-height: 1.15;
 }
 
 html,
