+++ examples/nextjs-example/app/head.tsx
@@ -1,0 +1,10 @@
+export default function Head() {
+  return (
+    <>
+      <title>Novu | NextJS Example</title>
+      <meta name="description" content="Generated by create next app" />
+      <meta name="viewport" content="width=device-width, initial-scale=1" />
+      <link rel="icon" href="/home/shane/Workspace/novu/apps/web/public/favicon.ico" />
+    </>
+  );
+}
