+++ examples/nextjs-example/app/page.module.css
