--- examples/nextjs-example/pages/_document.tsx
@@ -1,13 +1,0 @@
-import { Head, Html, Main, NextScript } from 'next/document';
-
-export default function Document() {
-  return (
-    <Html lang="en">
-      <Head />
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-}
